+++ src/main/java/br/com/etyllica/core/ui/UICoreListener.java
@@ -1,0 +1,5 @@
+package br.com.etyllica.core.ui;
+
+public interface UICoreListener {
+	public void changeApplication();
+}
