+++ src/main/java/examples/etyllica/gui/GuiExamples.java
@@ -30,9 +30,9 @@ public class GuiExamples extends EtyllicaApplet {
 		//Change Theme
 		//ThemeManager.getInstance().setTheme(new EtyllicTheme());
 		
-		//return new GeneralGuiExample(w,h);
+		return new GeneralGuiExample(w,h);
 		//return new TextFieldExample(w,h);
-		return new SliderExample(w,h);
+		//return new SliderExample(w,h);
 	}
 	
 }
